+++ .travis.yml
@@ -20,3 +20,4 @@ install:
 
 script:
   - npm test
+
