+++ tests/dummy/app/dynamic/controller.js
@@ -50,7 +50,7 @@ export default Ember.Controller.extend({
 
         contextMenuReportClicked: function(node, tree) {
             var self = this;
-            // this.set('lastItemClicked', '"Report" item for node: <' + node.text + '> was clicked.');
+            this.set('lastItemClicked', '"Report" item for node: <' + node.text + '> was clicked.');
         },
 
         addChildByText: function(nodeTextName) {
